[tune/structure] Introduce `stopper` package
